- Renames project to be more discoverable/accurate - Shrank summary to within 27 characters - Tweaked numbers display for four digit values
