+++ apps/apps.go
@@ -136,7 +136,7 @@ import (
 	"tidbyt.dev/community/apps/pathtrainschedule"
 	"tidbyt.dev/community/apps/petpikachu"
 	"tidbyt.dev/community/apps/phaseofmoon"
-	"tidbyt.dev/community/apps/plausiblestats"
+	"tidbyt.dev/community/apps/plausibleanalytics"
 	"tidbyt.dev/community/apps/pokedex"
 	"tidbyt.dev/community/apps/pollencount"
 	"tidbyt.dev/community/apps/powerball"
@@ -347,7 +347,7 @@ func GetManifests() []manifest.Manifest {
 		pathtrainschedule.New(),
 		petpikachu.New(),
 		phaseofmoon.New(),
-		plausiblestats.New(),
+		plausibleanalytics.New(),
 		pokedex.New(),
 		pollencount.New(),
 		powerball.New(),
