+++ apps/plausibleanalytics/plausibleanalytics.go
@@ -1,5 +1,5 @@
 // Package plausiblestats provides details for the Plausible Stats applet.
-package plausiblestats
+package plausibleanalytics
 
 import (
 	_ "embed"
@@ -7,19 +7,19 @@ import (
 	"tidbyt.dev/community/apps/manifest"
 )
 
-//go:embed plausible_stats.star
+//go:embed plausibleanalytics.star
 var source []byte
 
 // New creates a new instance of the Plausible Stats applet.
 func New() manifest.Manifest {
 	return manifest.Manifest{
-		ID:          "plausible-stats",
+		ID:          "plausible-analytics",
 		Name:        "Plausible Stats",
 		Author:      "brettohland",
-		Summary:     "Display your website's Plausible stats",
-		Desc:        "Works with your Plausible.io account to display your stats",
-		FileName:    "plausible_stats.star",
-		PackageName: "plausiblestats",
+		Summary:     "Plausible Analytics Display",
+		Desc:        "Display your website's traffic stats using your Plausible Analytics account.",
+		FileName:    "plausibleanalytics.star",
+		PackageName: "plausibleanalytics",
 		Source:  source,
 	}
 }
